refactor: PurchaseInfo 패키지 이동
