+++ src/main/java/lotto/view/ResultView.java
@@ -1,7 +1,7 @@
 package lotto.view;
 
 import lotto.domain.dto.LottoTicketDto;
-import lotto.domain.price.PurchaseInfo;
+import lotto.domain.ticket.PurchaseInfo;
 import lotto.domain.prize.LottoPrizeResult;
 import lotto.domain.prize.Prize;
